This version now has a list in room.py that hold the values of all room adjacent to each room
This version also has a list in the main file that creates a button with a given location and scale of the button size for each room. with this I have created a for loop tha will show all of the possible adjacent rooms the player currently in turn can move to.

Not all points have been plotted nor has the logic for allowing the player to click on these and move them to the selected room. right now each button is properly placed but the buttons will just move the to next room in numerical order.