added shuffling of the deck to the createDeck() def in the card.py file.
In the main file I created 4 lists that will store the hand of cards for each player in the game, I then made a list to keep all of these lists in.
I then made two functions, startingCards() and drawCard()

drawCard() will accept the index of the player, the list of all players hands and the deck of cards as parameters.
With these parameters the function will store the last card in the deck and the specficed players hand in variables.
The function will then add the last card to the players hand and delete that last card from the deck.

startingCards() is a work in progress but will loop through the list of every players hand.for each hand the function will call the darwCard() function and loop 6 times to give each player their starting set of 6 cards