added functionality that changed the action of continuing to the next players turn. Now users will have to click the next turn button to end their turn. I have also added an action point system
The Action point system gives each player a point for movement and a point for playing a card. every turn a player has the option to move one space, players are also able to play a card as well. So far we only have movement type cards and when a players uses a movement card the face value of the card will be added to the movement action points while also setting the movementAction variable will be set to True. After that the play can move for the amount of points they how have or also only partially use the points. If partially used the points will be set back to 1 when the next turn is started.

Now players can take the actions of either moving and or playing a movement card to gain movement points within their turn.