Can now Draw cards, reshuffle empty deck and Dr Lucky fixes
I first created a temporary draw card button that we will replace later. The purpose of this button is to allow for a player to draw a card manualy. When the button is pressed it will execute the drawcard function and set the cardActionPoints to 0 since a player can either draw a card or play one.

I have made the drawCard function now check if the main card deck is empty and if so it will shuffle the discard pile, set the main deck equal to the discard pile and clear the discard pile. After that the fucntion will draw a card for the player.

Lastly I made it so when the first cards are given out to each player that it skips doctor lucky. The startingCards function will check if the player is equal to dr lucky and skip the drawing if so. This logic is not needed in the drawCardButton logic because it will use the player value that the current turnorder is equal to and the turnorder will never be set to dr lucky.