added siteline/attacking functionality and fixed movement issues > > I first added a Kill button to the sidebar, this button will
check for a set of condtions such as if Dr. Lucky and the player are in the same room, if no players are in any of the rooms that are within site, and that the player and Dr. Lucky are the only players in the room.

To check that no players are in any of the rooms within site I made a for loop that will loop though each element of the siteline list for the given room. for each element it will check if there are any players and if so it will set the var canKill to false. The var is initially set to true so if any room has a player it will be set to false making it so the player cannot attack.

When testing the program I noticed that when a player goes to enter a room that dr lucky or a player just exited the entering player will be placed either on top of another player in the room if any or will have the next player to enter placed on top of them.

To fix this I edited the updatePlayer() function in the player class by removing the function of adding to the room count when the function is called. Now the room count will be updated in the main file when movement actions are triggered. After that I made a function in the main file that would repostion all of the player in a room whenever a player exits the room. so when a player exits a room you will notice that the players in that room will be repostioned. this is done by reseting the room count and looping through the player list, while looping it will check of the player is in the room that was exited and for each player that is it will repostion them.