Adjusted the card system to now allow players to play as many cards as they want in a turn.
The new rules I have implemented are as follows:
- players cannot draw a card after playing a card and players cannot play a card after drawing one
 - Players players cannot draw another card after drawing one

 - Players cannot play more than one weapon card

I have done all of this be removing the card action point system and creating these variables: WeaponCard, drawCardAction and playCard.

As players play or draw cards these values are appropriatly changed to true/ false

On top of this I also created a var called murderAttempt and set it to False be default. When a player atempts to kill Dr. Lucky it will change to True and if they attempt to attack again they cannot if they have alreadt tried during their turn.